services/purchasing-ai/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.0

# Forecasting
prophet==1.1.5
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# LLM
openai==1.14.0

# Utils
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.3
apscheduler==3.10.4
